docs(release-workflow): define “release definition”
I found it unclear when this article referred to some “package release definitions”—but only once, even though it seems like a new term. Googling it was tough, but I eventually landed on https://yarnpkg.com/getting-started/qa which stated that package release definitions are stored in the `.yarn/versions` directory. This is how I understood what “release definitions” discussed in *this* article are.